Lib: Logging PR Fixups
